+++ R/glmnet.R
@@ -22,6 +22,7 @@
 #' (lasso) versus L2 (ridge) regularization.
 #' - `alpha = 1`: Pure lasso model
 #' - `alpha = 0`: Pure ridge model
+#' @param ... Additional named arguments passed to [glmnet::glmnet()].
 #' @inheritParams glm_offset
 #'
 #' @returns A `glmnet` object. See [glmnet::glmnet()] for full details.
@@ -35,22 +36,35 @@
 #'
 #' @seealso [glmnet::glmnet()]
 #' @export
-glmnet_offset <- function(x, y, family, offset_col = "offset",
-                          weights = NULL, lambda = NULL, alpha = 1) {
-
+glmnet_offset <- function(
+  x,
+  y,
+  family,
+  offset_col = "offset",
+  weights = NULL,
+  lambda = NULL,
+  alpha = 1,
+  ...
+) {
   rlang::check_installed("glmnet")
 
   if (!offset_col %in% colnames(x)) {
-    rlang::abort(glue("A column named `{offset_col}` must be present."))
+    cli::cli_abort("A column named `{offset_col}` must be present.")
   }
 
   offsets <- x[, offset_col]
   x <- x[, !colnames(x) %in% offset_col, drop = FALSE]
 
-  glmnet::glmnet(x, y, family = family, weights, offset = offsets,
-                 lambda = lambda,
-                 alpha = alpha)
-
+  glmnet::glmnet(
+    x,
+    y,
+    family = family,
+    weights,
+    offset = offsets,
+    lambda = lambda,
+    alpha = alpha,
+    ...
+  )
 }
 
 # code from the parsnip package