+++ R/xgboost.R
@@ -23,35 +23,52 @@
 #'   us_deaths$off <- log(us_deaths$population)
 #'   x <- model.matrix(~ age_group + gender + off, us_deaths)[, -1]
 #'
-#'   mod <- xgb_train_offset(x, us_deaths$deaths, "off",
-#'                           eta = 1, colsample_bynode = 1,
-#'                           max_depth = 2, nrounds = 25,
-#'                           counts = FALSE)
+#'   mod <- xgb_train_offset(
+#'       x, us_deaths$deaths,
+#'       "off",
+#'       eta = 1,
+#'       colsample_bynode = 1,
+#'       max_depth = 2,
+#'       nrounds = 25,
+#'       counts = FALSE
+#'   )
 #'
 #'   xgb_predict_offset(mod, x, "off")
 #' }
 #'
 #' @export
 xgb_train_offset <- function(
-    x, y, offset_col = "offset", weights = NULL,
-    max_depth = 6, nrounds = 15, eta  = 0.3, colsample_bynode = NULL,
-    colsample_bytree = NULL, min_child_weight = 1, gamma = 0, subsample = 1,
-    validation = 0, early_stop = NULL, counts = TRUE, ...) {
-
-  rlang::check_installed("xgboost")
+  x,
+  y,
+  offset_col = "offset",
+  weights = NULL,
+  max_depth = 6,
+  nrounds = 15,
+  eta = 0.3,
+  colsample_bynode = NULL,
+  colsample_bytree = NULL,
+  min_child_weight = 1,
+  gamma = 0,
+  subsample = 1,
+  validation = 0,
+  early_stop = NULL,
+  counts = TRUE,
+  ...
+) {
+  rlang::check_installed("xgboost", version = "3.0.0")
 
   others <- list(...)
 
   if (!is.numeric(validation) || validation < 0 || validation >= 1) {
-    rlang::abort("`validation` should be on [0, 1).")
+    cli::cli_abort("`validation` should be on [0, 1).")
   }
 
   if (!is.null(early_stop)) {
     if (early_stop <= 1) {
-      rlang::abort(paste0("`early_stop` should be on [2, ",  nrounds, ")."))
+      cli::cli_abort("`early_stop` should be on [2, {nrounds}).")
     } else if (early_stop >= nrounds) {
       early_stop <- nrounds - 1
-      rlang::warn(paste0("`early_stop` was reduced to ", early_stop, "."))
+      cli::cli_warn("`early_stop` was reduced to {early_stop}.")
     }
   }
 
@@ -59,13 +76,16 @@ xgb_train_offset <- function(
   # subtract one column for the offset
   p <- ncol(x) - 1L
 
-  x <- as_xgb_data_offset(x, y, offset_col,
-                          validation = validation,
-                          weights = weights)
-
+  x <- as_xgb_data_offset(
+    x,
+    y,
+    offset_col,
+    validation = validation,
+    weights = weights
+  )
 
   if (!is.numeric(subsample) || subsample < 0 || subsample > 1) {
-    rlang::abort("`subsample` should be on [0, 1].")
+    cli::cli_abort("`subsample` should be on [0, 1].")
   }
 
   # initialize
@@ -81,9 +101,10 @@ xgb_train_offset <- function(
   }
 
   if (min_child_weight > n) {
-    msg <- paste0(min_child_weight, " samples were requested but there were ",
-                  n, " rows in the data. ", n, " will be used.")
-    rlang::warn(msg)
+    cli::cli_warn(paste0(
+      "{min_child_weight} samples were requested but there were ",
+      "{n} rows in the data. {n} will be used."
+    ))
     min_child_weight <- min(min_child_weight, n)
   }
 
@@ -94,19 +115,24 @@ xgb_train_offset <- function(
     colsample_bytree = colsample_bytree,
     colsample_bynode = colsample_bynode,
     min_child_weight = min(min_child_weight, n),
-    subsample = subsample
+    subsample = subsample,
+    objective = "count:poisson"
   )
+
+  if ("nthread" %in% names(others)) {
+    arg_list[["nthread"]] <- others$nthread
+    others[["nthread"]] <- NULL
+  }
 
   others <- process_others(others, arg_list)
 
   main_args <- c(
     list(
       data = quote(x$data),
-      watchlist = quote(x$watchlist),
+      evals = quote(x$evals),
       params = arg_list,
       nrounds = nrounds,
-      early_stopping_rounds = early_stop,
-      objective = "count:poisson"
+      early_stopping_rounds = early_stop
     ),
     others
   )
@@ -117,9 +143,13 @@ xgb_train_offset <- function(
 }
 
 # Original code from parsnip - modified for offsets
-as_xgb_data_offset <- function(x, y, offset_col,
-                               validation = 0, weights = NULL) {
-
+as_xgb_data_offset <- function(
+  x,
+  y,
+  offset_col,
+  validation = 0,
+  weights = NULL
+) {
   n <- nrow(x)
 
   if (is.data.frame(x)) {
@@ -127,14 +157,14 @@ as_xgb_data_offset <- function(x, y, offset_col,
   }
 
   if (!offset_col %in% colnames(x)) {
-    rlang::abort(glue("A column named `{offset_col}` must be present."))
+    cli::cli_abort("A column named `{offset_col}` must be present.")
   }
   offsets <- x[, offset_col]
   x <- x[, !colnames(x) %in% offset_col, drop = FALSE]
 
   # exit early for predictions where y's aren't supplied
   if (missing(y)) {
-    return(xgboost::xgb.DMatrix(x, info = list(base_margin = offsets)))
+    return(xgboost::xgb.DMatrix(x, base_margin = offsets))
   }
 
   if (!inherits(x, "xgb.DMatrix")) {
@@ -146,28 +176,26 @@ as_xgb_data_offset <- function(x, y, offset_col,
         data = x[-trn_index, , drop = FALSE],
         label = y[-trn_index],
         missing = NA,
-        info = list(base_margin = offsets[-trn_index])
+        base_margin = offsets[-trn_index]
       )
-      watch_list <- list(validation = val_data)
+      evals <- list(validation = val_data)
 
-      info_list <- list(label = y[trn_index],
-                        base_margin = offsets[trn_index])
-      if (!is.null(weights)) {
-        info_list$weight <- weights[trn_index]
-      }
       dat <- xgboost::xgb.DMatrix(
         data = x[trn_index, , drop = FALSE],
         missing = NA,
-        info = info_list
+        label = y[trn_index],
+        base_margin = offsets[trn_index],
+        weight = weights[trn_index]
       )
-
     } else {
-      info_list <- list(label = y, base_margin = offsets)
-      if (!is.null(weights)) {
-        info_list$weight <- weights
-      }
-      dat <- xgboost::xgb.DMatrix(x, missing = NA, info = info_list)
-      watch_list <- list(training = dat)
+      dat <- xgboost::xgb.DMatrix(
+        x,
+        missing = NA,
+        label = y,
+        base_margin = offsets,
+        weight = weights
+      )
+      evals <- list(training = dat)
     }
   } else {
     dat <- xgboost::setinfo(x, "label", y)
@@ -175,10 +203,10 @@ as_xgb_data_offset <- function(x, y, offset_col,
     if (!is.null(weights)) {
       dat <- xgboost::setinfo(x, "weight", weights)
     }
-    watch_list <- list(training = dat)
+    evals <- list(training = dat)
   }
 
-  list(data = dat, watchlist = watch_list)
+  list(data = dat, evals = evals)
 }
 
 # code from the parsnip package
@@ -198,27 +226,29 @@ recalc_param <- function(x, counts, denom) {
 # code from the parsnip package
 maybe_proportion <- function(x, nm) {
   if (x < 1) {
-    msg <- paste0(
-      "The option `counts = TRUE` was used but parameter `", nm,
-      "` was given as ", signif(x, 3), ". Please use a value >= 1 or use ",
-      "`counts = FALSE`."
-    )
-    rlang::abort(msg)
+    cli::cli_abort(c(
+      paste0(
+        "The option `counts = TRUE` was used but parameter `{nm}`",
+        " was given as {signif(x, 3)}."
+      ),
+      i = "Please use a value >= 1 or use `counts = FALSE`."
+    ))
   }
 }
 
 # code from the parsnip package with small edits
 process_others <- function(others, arg_list) {
-
-  guarded <- c("data", "weights", "watchlist", "objective")
+  guarded <- c("data", "weights", "evals", "objective")
   guarded_supplied <- names(others)[names(others) %in% guarded]
 
   n <- length(guarded_supplied)
   if (n > 0) {
-    rlang::warn(
-      c("!" = glue(
-        "The following arguments are guarded and will not be passed to ",
-        "`xgb.train()`: {paste(guarded_supplied, collapse = ', ')}.")
+    cli::cli_warn(
+      c(
+        "!" = paste0(
+          "The following arguments are guarded and will not be passed to ",
+          "`xgb.train()`: {guarded_supplied}."
+        )
       ),
       class = "xgboost_guarded_warning"
     )
@@ -227,10 +257,13 @@ process_others <- function(others, arg_list) {
   others <- others[!(names(others) %in% guarded)]
 
   if (!is.null(others$params)) {
-    rlang::warn(
-      c("!" = paste0("Please supply elements of the `params` list argument as ",
-                     "main arguments to `set_engine()` rather than as part of ",
-                     "`params`.")
+    cli::cli_warn(
+      c(
+        "!" = paste0(
+          "Please supply elements of the `params` list argument as ",
+          "main arguments to `set_engine()` rather than as part of ",
+          "`params`."
+        )
       ),
       class = "xgboost_params_warning"
     )
@@ -249,17 +282,17 @@ process_others <- function(others, arg_list) {
 #' @rdname xgb_train_offset
 #' @export
 xgb_predict_offset <- function(object, new_data, offset_col = "offset", ...) {
-
   if (!inherits(new_data, "xgb.DMatrix")) {
     new_data <- maybe_matrix(new_data) |>
       as_xgb_data_offset(offset_col = offset_col)
   } else {
     if (is.null(xgboost::getinfo(new_data, "base_margin"))) {
-      rlang::abort(paste0("If `new_data` is an `xgb.DMatrix`, it must have an ",
-                          "offset (base_margin) defined."))
+      cli::cli_abort(paste0(
+        "If `new_data` is an `xgb.DMatrix`, it must have an ",
+        "offset (base_margin) defined."
+      ))
     }
   }
 
   stats::predict(object, new_data, ...)
-
 }